requirements.txt
django
pillow
djangorestframework
django-filter